pyproject.toml
[tool.poetry]
name = "rho_plus"
version = "0.3.0"
description = "Aesthetic and ergonomic enhancements to common Python data science tools"
authors = ["Nicholas Miklaucic <nicholas.miklaucic@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8, <3.11"
scipy = ">=1.7.0"

[tool.poetry.dev-dependencies]
matplotlib = "^3"
scipy = ">=1.7.0"
numpy = "^1.21"
colour-science = "^0.4"
seaborn = "^0.11"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
